backend/requirements.txt
google-adk==1.26.0
google-cloud-firestore==2.23.0
google-genai==1.65.0
fastapi==0.135.0
uvicorn==0.41.0
websockets==15.0.1
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.13.1
